Created master.css with h1 and linked it to index.html
In the index.html, I eliminated the style tags and replaced it with a link to a separate master.css file. The link has appropriate rel, type, and media tags in addition to href tag to the css file. In the master.css, I changed the color of h1 to red to test it out.