Removing target="_blank" from <base>
Having target="_blank" in the <base> of <head> made my anchor tags automatically open in a new tab. If you want to make opening in a tab the default in the future, add target="_self" to the same page anchor tags and remove target="_blank" from the open in new tab links.